правки на основании комментариев наставника
