+++ src/utils/filter.js
@@ -1,7 +1,7 @@
 const moviesToFilterMap = {
   all: (movies) => movies.length,
   watchlist: (movies) => movies
-    .filter((movie) => movie.isInWatchlist).length,
+  .filter((movie) => movie.isInWatchlist).length,
   favorites: (movies) => movies
   .filter((movie) => movie.isInFavorites).length,
   history: (movies) => movies
